+++ src/webapp.test.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import ReactDom from 'react-dom';
-
-describe('e2e Padachone', () => {
-    it('works fine', () => {
-        expect(true).toBe(true);
-        
-    })
-})